permisions

+++ src/app/services/logs-alerts.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@angular/core";
 import { HttpClient,HttpHeaders  } from "@angular/common/http";
 import { LogsAlert } from "../models/logs_alert.model";
+import { PermsService } from "./perms.service";
 
 @Injectable({
   providedIn: 'root'
@@ -8,7 +9,7 @@ import { LogsAlert } from "../models/logs_alert.model";
 export class LogsAlertsService {
   readonly URL_API = "http://localhost:3000/api/log_alerts";
   token = localStorage.getItem('token');
-  constructor(private http: HttpClient) {
+  constructor(private http: HttpClient, private permsService: PermsService) {
   }
 
   
@@ -19,25 +20,55 @@ export class LogsAlertsService {
   };
 
 
-  postLogsAlert(LogsAlert: LogsAlert) {
-    return this.http.post<LogsAlert>(this.URL_API, LogsAlert,this.httpOptions);
+  async postLogsAlert(LogsAlert: LogsAlert) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'create_logs');
+    if (perms) {
+      return this.http.post<LogsAlert>(this.URL_API, LogsAlert,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
-  getLogsAlerts() {
-    return this.http.get<LogsAlert[]>(this.URL_API+``,this.httpOptions);
+  async getLogsAlerts() {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'read_logs');
+    if (perms) {
+      return this.http.get<LogsAlert[]>(this.URL_API+``,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
 
-  getLogsAlert(id:string) {
-    return this.http.get<LogsAlert>(this.URL_API+`/${id}`,this.httpOptions);
+  async getLogsAlert(id:string) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'read_logs');
+    if (perms) {
+      return this.http.get<LogsAlert>(this.URL_API+`/${id}`,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
-  putLogsAlert(id:string, LogsAlert: LogsAlert) {
-    return this.http.put<LogsAlert>(this.URL_API+`/${id}`,LogsAlert,this.httpOptions);
+  async putLogsAlert(id:string, LogsAlert: LogsAlert) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'edit_logs');
+    if (perms) {
+      return this.http.put<LogsAlert>(this.URL_API+`/${id}`,LogsAlert,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
-  deleteLogsAlert(id: string) {
-    return this.http.delete(this.URL_API+`/${id}`,this.httpOptions);
+  async deleteLogsAlert(id: string) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'delete_logs');
+    if (perms) {
+      return this.http.delete(this.URL_API+`/${id}`,this.httpOptions);
+    } else {
+      return null;
+    }
   }
