+++ src/app/services/product.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@angular/core";
 import { HttpClient,HttpHeaders  } from "@angular/common/http";
 import { Product } from "../models/product.model";
+import { PermsService } from "./perms.service";
 
 
 
@@ -10,7 +11,7 @@ import { Product } from "../models/product.model";
 export class ProductService {
   readonly URL_API = "http://localhost:3000/api/product";
   token = localStorage.getItem('token');
-  constructor(private http: HttpClient) {
+  constructor(private http: HttpClient, private permsService: PermsService) {
   }
 
   
@@ -21,26 +22,67 @@ export class ProductService {
   };
 
 
-  postProduct(Product: Product) {
+  async postProduct(Product: Product) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'create_product');
+    if (perms) {
+      return this.http.post<Product>(this.URL_API, Product,this.httpOptions);
+    } else {
+      return null;
+    }
     return this.http.post<Product>(this.URL_API, Product,this.httpOptions);
   }
 
-  getProducts() {
-    return this.http.get<Product[]>(this.URL_API+``,this.httpOptions);
+  async getProducts() {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'read_product');
+    if (perms) {
+      return this.http.get<Product[]>(this.URL_API+``,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
 
-  getProduct(id:string) {
-    return this.http.get<Product>(this.URL_API+`/${id}`,this.httpOptions);
+  async getProduct(id:string) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'read_product');
+    if (perms) {
+      return this.http.get<Product>(this.URL_API+`/${id}`,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
-  putProduct(id:string, Product: Product) {
-    return this.http.put(this.URL_API+`/${id}`,Product,this.httpOptions);
+  async getProductByCode(id:string) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'read_product');
+    if (perms) {
+      return this.http.get<Product>(this.URL_API+`/code/${id}`,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
-  deleteProduct(id: string | undefined) {
+  async putProduct(id:string, Product: Product) {
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'edit_product');
+    if (perms) {
+      return this.http.put(this.URL_API+`/${id}`,Product,this.httpOptions);
+    } else {
+      return null;
+    }
+  }
+
+  async deleteProduct(id: string | undefined) {
     if (!id) throw new Error('ID is required');
-    return this.http.delete(this.URL_API+`/${id}`,this.httpOptions);
+    const id_user = localStorage.getItem('id');
+    const perms = await this.permsService.getPerms(id_user, 'delete_product');
+    if (perms) {
+      return this.http.delete(this.URL_API+`/${id}`,this.httpOptions);
+    } else {
+      return null;
+    }
   }
 
 }
